Move JSON-parsing into own module; replace JsonData with Validator
Validator has a much simpler interface, dispensing with the context manager functionality and returning a new instance from the index operator. That makes it much easier to statically type Validator. It also simplifies list processing. And the result of the index operator is consistent with developer expectations. In short, a much improved wrapper class.